+++ semana-5/introducao-ao-react/labetube-vanilla/.gitignore
